Prepare for the week ahead.
Last week was just generally not good in the educational department.
I didn't get good sleep and I barely managed to get things in on time.
This week, we're gonna try to make it a little different. If I could
hammer out all of this to completely focus on math. That'd be good.